Assets Added + Build Instructions & GIF in README
Added in the assets that weren't uploaded before for some strange reason.

Added short build instructions to the README, as well as a couple of gifs. Unfortunately I think trying to set up a system that fetches glfw and installs it will take me having to reconfigure my whole project to work with CMake, which sounds like a pain to me. So this is a subpar work around.